Added Error handling
